ACTIN-37 Add JT interval and evaluation in cardiac functions (#15)
Add JT interval and evaluation in cardiac functions

In adding the new function the duplication between the QTCF limit and sufficiency functions
is refactored into a shared class with configurable threshold behaviour (min/max). A 
simple static factory class encapsulates the 3 usages (min/max qtcf, min jtc)
The shared class removes the need for the shared QTCFunctions and introduces a new type
ECGMeasure to combine the value and unit, and make null checking more atomic.

The new clinical inputs are added to the curation config and model.

The inputs are also persisted in the clinicalStatus table.