ACTIN-2 Derive tumor stage and apply across applicable evaluation functions
Create a function to derive the tumor stage based on precense of lesions. The
rules are modeled as orthogonal predicates, each mapping to a list of stages.

Create a general function DerivedTumorStageEvaluationFunction to encapsulate the
logic of evaluating the derived stages against an evaluation function, and merging
the Evaluation into a single status and merged messages/events.

HasTumorStage deviates from the standard handling of the DerivedTumorStageEvaluationFunction
so its handling of the derivations is in the function itself.