ACTIN-2 Combine the first two approaches
Keeps the factory based evaluation creation over the accumulation, but internally concatenates
all messages to the single message to preserve all information in all cases.

Display name field is introduced to Evaluation to be used for creation of the general rules,
as it's not possible to extract that information from the EvaluationFunction other than from
the class name (which could be a viable alternative)

Some messages within the functions are restructured to always include the tumor stage. And the
metastatic messages are fixed as they erroneously mention locally advanced cancer.