ACTIN-14 Add complication category to various prior conditions
Added to eye disease, cardiac disease, and stroke.

Has complication category is OR'd with has prior condition in the other
condition rule mapper.

Assume the "toxicity" reference in README for eye disease was a typo,
removed.

Feels weird not to have a test for this, but from what I can see the
that composite functions created in Mapper classes are generally not
tested. Would be up for adding some, but also don't want to introduce
unnecessary complexity or inconsistency. Could be that some kind of
higher level component/functional test makes more sense.