ACTIN-2 Derive tumor stage and apply across applicable evaluation functions
Create a function to derive the tumor stage based on presense of lesions. The
rules are modeled as orthogonal predicates, each mapping to a list of stages.

Create a general function DerivedTumorStageEvaluationFunction to encapsulate the
logic of evaluating the derived stages against an evaluation function, and merging
the Evaluation into a single status and merged messages/events.

Messages are merged into a single message associated to the derived status, where
the worst outcome of the derivation is used, and a note is attached to indicate which
stages were used for the evaluation.

HasTumorStage deviates from the standard handling of the DerivedTumorStageEvaluationFunction
so its handling of the derivations is in the function itself.