ACTIN-51 Use optionalArray when extracting xref
From both graph and metadata, as it can in practice be null. The null
check already existed.

Removed metadata xref from the minimal owl doid in the test example_doid.json
to ensure this case was covered by some test. Seem appropriate as the presence
of the xref is tested in node2.