ACTIN-58 Handle ambiguous complications as per spec (WIP)
As per the ticket, adds a special "empty" complication which downstream leads
to undetermined results in has complication rules.

I don't like this solution both for the magic empty strings/lists and that we
kind of duplicate the "maybe has a condition" between the boolean in the
clinical input and the list of complications.

But it should work, might be worth trying out for patient 201 and then improving
along with ACTIN-1.