ACTIN-58 Handle ambiguous complications as per spec (WIP)  (#25)
As per the ticket, adds a special "empty" complication curation. The empty curation will be mapped to
input like "yes". Also a boolean is added as convenience 

Downstream the empty "yes input" is used in the undetermined status of complication string matching 
(both on name and category). Has any complication will return PASS when this field is set.

The yes input is filtered from DB loading, and a new column is added to the DB to persist the new 
boolean.

Overall there lots to improve here. We should make this more explicit and deliberate. This will be 
addressed in ACTIN-1.