ACTIN-84 Integrate SERVE known genes with ACTIN
Use the new KnownGene datamodel from SERVE and integrate in with ORANGE
interpretation and treatment creation.

In ORANGE we remove the known genes tsv argument and all references,
preferring to use the collection exposed in KnownEvents.

In treatment creator we still load the file directly. Debatable, as
maybe we should just load the entire serve kb, but then we get more
than we need and require a ref genome argument.