ACTIN-173 Extract dosage, frequency and period information from clinical feed
Previously was completely dependent on curation to populate dosage fields.

Extracts dosage into its own class and takes values from the feed, unless the administration is oral and 
dosage information is missing (then reverts back to curation). 

Period unit and dosage unit are curated/translated independently, as they can still be incorrect or in NL
in the feed.

The medication table in report is also updated to add more medication information.