ACTIN-188 Expose ATC codes via clinical record and check feed for validity (#109)
Adds the ATC classification at clinical ingestion based on the code taken from feed. The feed also includes
its own values for the levels, and these are checked against ATC tree for validity. If they don't match we
warn the user and accept the entry.

The datamodel changes such that the ATC classification are stored in an object, with the full level breakdown. 
Each level keeps its code and name for use in rules. 

The ATC codes themselves will be ingested from a TSV representation of the tree data originating from WHO.

NOTE: As we do not yet have the ATC tree from WHO, this feature is disables in this PR.