ACTIN-262 Enable useage of full ATC tree
The structure/field names in the real ATC tree are a little different from the
sample. It is also possible to have duplicates of the same medication with different
dosage information.

A test is added to ensure the real format is parseable.

The WhoAtcModel is now used instead of the noop.