ACTIN-267 Enable useage of full ATC tree (#146)
The structure/field names in the real ATC tree are a little different from the
sample. It is also possible to have duplicates of the same medication with different
dosage information.

A test is added to ensure the real format is parseable.

Make the 5th level of ATC classification optional, and warn/return null if any
further levels are missing.

Remove the validation warnings, we've seen that the data is in line with the
ATC tree values.