ACTIN-301: Add treatment intent to report
We model multiple intents per treatment history entry.
These will now be included in lower-case and space
separated in a new portion of the combined annotation.

The Palliative intent is excluded.

Note that the order of the intent entries (when there are
multiple in a history entry) is the iteration order
of the (immutable) set, which presumably matches the
data from the upstream source json.