ACTIN-411 Expose uncaught exceptions to the log in a predictable way (#233)
Tried first to expand the results.json to include more failure modes, but the
ways things can fail are spread out between the parsing of inputs, construction
of models, and the execution of models themselves.

So to be comprehensive instead ensured that we actually catch all exceptions
and predictably log the error message (and exit 1). Pipelines can then properly
handle the failure by presenting the unexpected ERROR to the user.