ACTIN-394 Expose trial definition and ctc database ingestion warnings (#222)
Refactors the trial ingestion to expose the warnings expressed by the logger as validation errors. These errors are
persisted to a results json file at the end of the run and can be used by downstream applications. They are also 
logged as warnings at the end of the run, rather than during.