ACTIN-433 Validate curation input  (#249)
Move from throwing runtime exceptions to capturing all problems in the curation input and persisting them to a json file for downstream processes. If any validation errors are encountered, the ingestion is halted.

As part of this functionality we made a large refactoring to decompose the curation database into their individual 
components, and have the extractors only depend on the components they use.

Tests have been added for all the config factories.

The ClinicalIngestionTest now serves as a component test, loading realistic feed and curation files from the test resources.