ACTIN-481 Fix issues with reporting of unused configs
Compare against the keys of the maps (which have been lowercased) instead of the input
in the config/translation.

Added an example unused warning to ClinicalIngestionTest for each of curation databases
and translations. All other unused configs are removed from the test curation files,
for now. May add them again to make the test more comprehensive but that is not in scope
of this change.