ACTIN-498 Replace our own TSV parser with Jackson
Jackson's CSV/TSV mappers are pretty mature and feature rich, as you can see we can remove a lot of code and
probably have a more robust, bug free implementation.

This is just a quick spike, the only working feed entry is the PatientEntry.  We'd need to add some custom
deserializers for the special values.

We'd also need a place for the isValid method, but from a quick scan most of those are trivial and probably
make sense to add to our other validation errors (curation, etc).

Curious about your thoughts!