ACTIN-525 New clinical feed model for NKI/Kaiko feed (#320)
Creates a new module and model to extract the clinical feed data from a new model developed together with Kaiko,
based on the specifications in the clinical readme.

The extraction leverages existing curation and warning propagation in the existing feed extractor. That said, I've tried to
aggressively reduce the amount of curation we need to do. The Kaiko data should be better structured than what we've
seen previously. We do however have some types that require individual curation.

The different feed sources can be toggled as an argument to the ClinicalIngestionApplication.