ACTIN-593 Implementation ready for draft review
The test cases annotated with `@Ignore` do not pass as-written. I believe they
are correct given the requested functionality (see JIRA) but the provided
implementation follows a more-accepted pattern based on other rules I already
see which makes me wonder whether the specification might have left out some
details which are assumed to be known.

Specifically referring to the use of the `TreatmentSummaryForCategory` class
and the `matchesTypeFromSet` method which invokes the concept of a uncertainty
based on trial participation.