ACTIN-593 Reimplement
The logic was just too squirrely with the use of the helper classes,
I do not think they were really designed for this use case.

Underpinning the decision to change was that the logic for determining
the difference between whether the named treatment and the "other" with
the same category and optionally types were one and the same was way too
complicated once I got down to it.