ACTIN-692 Fallback to the original treatment when curation is not possible for modications
We've seen in practice that modification names are frequently not treatments, but more of a
"note" about a change in the current treatment.

This change will make it such that treatment extraction will use the original treatment in
the case of a modification with an uncurateable name.

When the treatment is not ignored, a warning will be returned for the curation sheet. If
ignored, the modification will use the treatment of the parent history entry and extract
the other modification fields (cycles/date)