ACTIN-478 Subjects sanitised on the way in
Don't store the unsanitised subjects from the feed, strip the
hyphens out before processing, in the CSV parsing layer.

While I did add annotations to a number of classes that did not
previously have them I do not think this is such a big problem
as some `Entry` implementations had each field annotated for
Jackson while others didn't.

Add a custom deserialiser for the subject fields and attach
it to every `Entry`. A `JsonProperty` annotation must be added
so that Jackson will pick up the custom deserialiser, when that
has happened all the fields must be annotated or Jackson will
throw when it is run.

Related but not strictly part of the ticket: constant for body weight
unit.