ACTIN-713 Ensure feed is of latest version
The Questionnaire is unit-tested for each version. This test is using
whichever version of the questionnaire is implemented by the files it
reads from the feed directory, and because we continue to support older
versions of the feed there is nothing to keep this test from testing
an old version of the feed. It makes most sense to test the latest
version of the feed since that is going to be what we are expecting
to receive in new patients at any given time.