ACTIN-715 Move FeedRecord back to its own file
Also update to have a companion object for performing the construction of
it and make it into a data class.

I moved the VitalFunctionProperties data class into FeedRecord's file for
expediency as that seemed better than increasing its visibility.

Open to suggestions on whether there should be a standalone test for the
FeedRecord, on the one hand I think there should be as it would target
the functionality of the companion object factory method, on the other
it does not add anything in the way of test coverage improvements.