ACTIN-725 Handle the null case for medications (#449)
Makes medications nullable on the clinical record and modifies all rules to handle that case.

Checks the medication for null and assigns to a variable or returns a consistent evaluation result. 

Also adds the scrubbing for NKI case, and handles the null in the report by excluding the
medication section entirely.