ACTIN-759 Allow toggling toxicity source warning
Introduce a configuration toggle to allow us to specify whether the
existing "n.b. different EHR source than questionnaire" note that goes
along with a warning instead of a pass, when the questionnaire is not
the source of toxicity data.

This is required for better readability of NKI reports. We don't have a
questionnaire (yet?) so the whole warning is not useful there.