ACTIN-759 Switch to ResourceLocator everywhere
There were inspection warnings when using the alternative that we had
adopted.

Also remove a redundant AlgoConfiguration parameter as pointed out in
the PR comments and make the object parameter optional on the
`ResourceLocator.onClasspath` method, defaulting to the `ResourceLocator`
object instance itself, thus allowing callers to use the method outside
of a class if they are happy to use the same classloader as the one
containing the `ResourceLocator` instance.

I also tidied up the `FeedModelTest` to use the AssertJ equivalents to
the JUnit originals.