ACTIN-759 Always use no-argument ResourceLocator
I originally added the `this` parameter so that we could guarantee that
the call would use the same classloader as would be used by the class if
it were doing its own calls but of course the distinction is not real
because the classloader in use will already be the one for the test that
is calling the locator method.

Remove all references to the parameter to prove the point.