ACTIN-759 Allow toggling toxicity source warning (#457)
Introduce a configuration toggle to allow us to specify whether the
existing "n.b. different EHR source than questionnaire" note that goes
along with a warning instead of a pass, when the questionnaire is not
the source of toxicity data.

This is required for better readability of NKI reports. We don't have a
questionnaire (yet?) so the whole warning is not useful there. May
revisit this in the near future to eliminate the need for the configuration
altogether but keeping it for now due to timelines.

Part of this was adding a static method for locating resources from tests
which eliminates the warnings associated with the Google class we
were using instead, so I changed the other tests to no longer use the
deprecated method call.

Further noticed the FeedModelTest was using the old-style Junit assertions
rather than the AssertJ ones we use everywhere else so changed them.