ACTIN-713 Questionnaire v1.7 for adapter test (#418)
The Questionnaire is unit-tested for each version. This test is using
whichever version of the questionnaire is implemented by the files it
reads from the feed directory, and because we continue to support older
versions of the feed there is nothing to keep this test from testing
an old version of the feed. It makes most sense to test the latest
version of the feed since that is going to be what we are expecting
to receive in new patients at any given time.

Therefore assertions have been added to make sure the version of 
the questionnaire being tested matches the latest available version and
a new test case added to guard against regression in the JSON.