ACTIN-754 PR suggestions + fix bug
The stage derivation was happening immediately after the
population of the tumor details which meant that there was
no information on the lesions (required for the derivation).

Move the derivation afterward to resolve this problem.

Also fix up formatting that had been identified in the PRs
and modify some of the existing test names.