ACTIN-878 Add report display for non-WGS non-IHC test types (#516)
Refactors the interpretation to be more general. Each "intepretation" represents various results for a given test type.
Within the results we have a grouping and details. The grouping can be either a common result (eg. Negative) or a
common gene (eg. EGFR) depending on the test type.

Then for display all types get one line per grouping, with comma delimited details. 

Adds a visitor to avoid a chain of instance-ofs, this can be used for other functionality involving aggregating results
of various types of molecular tests.