ACTIN-910 Add configuration to ignore new trial validation errors
When a new trial exists in the status database but not in the trial config.

Implemented as a configuration option as the error suppression in the pipelines
would not be feasible without adding a new validation error type to separate
from other status database validation errors.

Perhaps the changes for ACTIN-894 should be refactored to this approach if we
continue to support that optionality.