ACTIN-920 Fix reporting issue with driver type (#549)
During copy number extraction, the copy number is type is extracted from the purple somatic gains/losses
which only contain entries for driver copy numbers meeting the criteria of ploidy*3. For copy numbers
which we interpret based only on it being within a known gene, this lookup is not possible. For these
copy numbers we extract a status of None. This behaviour was added in ACTIN-45 when we began to look at
all purple copy numbers rather thank just the gains/losses.

Downstream, the reporting code was only looking at the isGain flag, if true we use the string Gain, if
false Loss. This causes the erroneous Loss in the report for a copy number 9. The fix here is just to the reporting code,
we use the string "Copy Number" when the type is None.