ACTIN-928 Match derived stages to sub-stages (#552)
If a derived stage has matching sub-stages (defined as a stage with the derived as category), then return
pass, as we have no logic to derive the sub-stage. For example, a derived stage of III will match IIIa.

This does not apply to non-derived stages, so if we know the stage is III we do not match it to IIIa.