ACTIN-945 Support annotation of non-orange molecular tests (#561)
Creates a some common structure around the extraction and annotation of various types of molecular tests with
the introduction of the MolecularExtractor and MolecularAnnotator interfaces. The existing orange interpretation is
retrofitted with these interfaces.

The MolecularTest class is broken up into several extractors. To make creating and passing around MolecularTests simpler, we change the MolecularTest interface to remove the generic result, and make the old result classes the implementations of MolecularTest.

An ArcherAnnotator is introduced but is not functionally complete as we wait on more comprehensive extraction of
coordinate and variant info. For now it does the following:

Tries to locate ActionableEvidence.
Tries to locate any known gene alterations
Tries to locate any known codons or exons, to add the rank as annotation
To support this the EvidenceDatabase (and supporting classes) are modified to take "criteria" data classes as opposed to
Variant itself. The criteria contain only properties required for matching, in a flattened format. Criteria can be easily constructed from MolecularRecord variants and ArcherVariants.