ACTIN-955 Cleaning up the clinical model use in NKI (and intended for future hospitals) (#572)
Reviewed the the current clinical model and revised with the following goals.

Remove all unused code and fields.

Improve naming. At some point I named everything EhrXxx but this isn't really accurate as not all data comes from an EHR. For instance, molecular tests could come from PALGA or another DB. Revised to use the term ProvidedXxx instead. Curious to hear opinions on this. External was another option we can consider.

Updated and moved the documentation of fields into the code, and generate a JSON schema which can be used to validate JSON inputs. While it's not great for human reading (although there are some tools that can generate MD from json schema), this is better for most data engineering teams, as it can be used programmatically check inputs. Also it will be easier to keep the docs up to date if it always matches the code. For now, generating the schema on-demand and it must be committed. We could add a git hook for this later.

Worth noting that moving over to FHIR is still a likely option in the long term, but still felt this was worthwhile to do now that the NKI clinical ingestion is stable.