ACTIN-971 Normalize variants and fusions between WGS molecular record and panel (#574)
Start sharing common interfaces for drivers, in particular variants and fusions, between the new panel extractions and the existing orange extraction.

Common interfaces are created to facilitate this. They contain everything which can be populated for both cases. In the case of the Drivers and Characteristics interfaces that doesn't 100% hold, but in the former case we can just have empty lists for drivers we cannot extract from the panels, and the latter already had all fields set as nullable.

Many of these fields are not yet populated. I expect this to happen in the XxxAnnotator classes. This will occur in subsequent tickets/prs. 

The panel data pipeline will now end in a common PanelRecord rather than the ArcherPanel and GenericPanel we had previously. This should be a more future proof approach as we add more panel types.

Note I did carry forward the old "panels" (now called XxxPanelExtractions) for backwards compatibility with existing working rules.

The "reference rule" for how to go forward is GeneHasActivatingMutation. We can now share the same logic for the PanelRecord as MolecularRecord.

I shuffled the packages around to make it more logical and cohesive. The new orange package contains all molecular 
datamodel specific to the interpretation of orange records.