ACTIN-1000 Enable special floating point numbers on GSON builder
When running ACTIN on historic patients, serialization of the molecular test history can fail on NaN
values (seen in ACTN01020134). From there it is not possible to deduce what field is NaN, as it's not
in the stack trace and the json is not persisted.

This change will allow the NaN to be persisted and followed up. Annoyingly, in the case of ACTN01020134,
there is no NaN in the output json after enabling this flag.