ACTIN-1019 Treat SUSPENDED trials as CLOSED at NKI
Without this update our system will require a manual override if a trial
is suspended in the status feed but enabled otherwise.

Roll the NKIStatusResolver functionality into the NKIStatusEntryReader
as I do not believe the abstraction was accomplishing what it may
originally have done, I do not feel it was helping.

Update the existing test cases for the trial status reading to include
the new SUSPENDED status.

Another part of this is that the "uninterpretable" status is gone, but I
cannot see how it ever would have been returned anyway since our filter
beforehand was already removing any unknown statuses.