ACTIN-996 Consolidate merging of results of multiple molecular tests (#593)
Consolidates the precedences of evaluations in a central class for molecular functions, using both the results and the
type of test.

After some back and forth, the decision was made to prefer WGS/NGS panel results from orange if we have multiple
evaluations as the same level (ie one WGS and one Archer, both passing, we take only the WGS messages and molecular events). If we have multiple orange results, we take the most recent. If we have multiple panels, we accumulate and merge the results in a single evaluation (ie one Archer and one AvL panel, events are merged from both evaluations).

I would have liked to move the consolidation out of the rules and into the shared MolecularEvaluationFunction. But this was
difficult due to the handling of undetermined in all the impacted rules. We kind of need the notion of a "PerGeneMolecularFunction" where you could make some determination of whether a gene was tested up front. I feel that's better done later when everything is using the MolecularTest interface and not digging into the extractions.