ACTIN-1303 Remove commented code
I do not understand how this was continuing to build all along, somehow
when I added the trial API dependency it "triggered" a failure in the
build related to this block. It builds fine without it for me so I am
just removing it and will deal with any fallout, I suppose it may only
appear at runtime.

The issue was that the `immutables.version` variable simply isn't
defined, I thought maybe it was something to do with the parent project
but that doesn't seem to be the case.