ACTIN-1307 Make cohortId optional
Required as oftentimes the cohorts coming in from the sheet will have
the MD5 (external id) filled in but not the cohortId (internal id). If
this situation occurs there would be an exception but with this change
there will just be warnings.

We can easily also filter out the ones with a null cohortId using this
change too which will be more compatible with ACTIN-1303.