ACTIN-1324 Add ECG curation to clinical status extraction
Simple prior other condition "cross curation". So we don't propagate the warnings
such that the curator does not need to ignore every prior other condition.

One question is what to do with the `hasSigAberrationLatestECG` which for EMC is
part of the questionnaire and thus always set. For now I've added is as a new
optional argument to curation, but curious to hear what you think!