ACTIN-1376 VAF not correctly used in molecular history (#802)
Two issues, the way we were structuring the test results meant we always used the VAF of the driver
which we were currently looping over for all tests.

Second, since VAF is part of the equality of a variant, we end up with rows drivers * tests and each
test row has 3 times the same VAF.

For now, fixed by a little presentation filtering and transforming. But we should revisit and probably
move VAF out of the variant interface.