ACTIN-1373: Wrap the existing system report examples in a regression test (#816)
Add tests that call the local reporter application with an example. Then assert whether all text matches,
which will catch any content issues. Then assert that the PDFs visually match, which will catch any
additional formatting problems.

The visual test leverages the iText compare tool. It turns everything to images then gives diffs per page
as images with the differences highlighted. This tool uses two external tools: ghostscript and image magick.
So we add a docker container with these tools to use as a build agent.

The workflow for a reporting change is this:
- These tests will be expected to fail after the report is changed.
- Developer should check the differences match up with the change they intended to make and nothing else.
- If something else changes, fix it.
- When only the intended change, we simply copy the new report into the examples directory and the test passes.