ACTIN-1754 Add locations to trials in driver table
Include the locations along with the acronym in the trial column in driver table.

This is preferable to additional columns per location, as we can have as many as 6
locations in current config (DRUP). Also keep in mind that LKO is not a location,
but a trial source and can also create trials in many locations beyond just EMC and
NKI.

Attached is a screenshot of the result.