ACTIN-1755 Put cohorts once in prefix only (#1009)
When we display the cohorts for trials in the report, they will all have
the same locations and thus displaying them for every row is redundant.

Move the location display into the "prefix" row and don't show them in
the rest of the table.

Also strongarm the locale for both the report regression test and the local
report wrapper application to "en-US" to avoid petty differences depending
on where it is run.