ACTIN-1763 Link to ct.gov for non-local trials (#1011)
When the requesting hospital is not the same as the source of the trial,
create a link to clinicaltrials.gov using the NCT id from the trial, or
if that is not present do nothing.

Re-uses the `includeLocation` boolean that was already present and which
previously only controlled the display of the hospital column; I think
they are close enough in intent that the case can be made to leave it
this way.