ACTIN-1755 Handle case with no prefix row
First time around with this ticket I modified the report to print the
locations only once, in the "prefix". Unfortunately this prefix area is
not always printed so sometimes we are now seeing no locations at all.

Detect the situation in which the prefix row is not printed and instead
put the locations in the row of the first cohort for the trial. I think
this could be a bit confusing visually but will be tidier than
reproducing the locations on every row.