ACTIN-1755 Put location in prefix row if >1 cohort
The previous implementation I think would have been confusing because it
would be ambiguous to a reader of the report whether all cohorts would
share the same location (something we know to be true but users may not)
or if only the first cohort had the given locations.

Remove the ambiguity by creating the prefix row even if none of the
other conditions hold, if there are more than a single cohort for the
trial.