ACTIN-1755 Rework display of locations (#1013)
First time around with this ticket I modified the report to print the
locations only once, in the "prefix". Unfortunately this prefix area is
not always printed so sometimes we are now seeing no locations at all.

In addition to the situations in which the prefix row would have been
printed originally (common data in two or more rows), print it if the location
display is on and there are more than two cohorts. These will all have the
same locations as that is specified at the trial level, so this is a suitable
situation to have the prefix too.