ACTIN-1745: Implement HasOnlyHadTreatmentWithCategoryOfTypes (#1005)
* Implement HasOnlyHadTreatmentWithCategoryOfTypes.kt and test class
* ACTIN-1745: Add rule to README, EligibilityRule and treatmentRuleMapper


Proposed logics:

PASS: The patient has had one or more treatments that all are of category X and have at least any type Y
FAIL: No treatments, or 1 or more treatments that don’t contain category X and/or type Y
WARN: The patient has one or more treatments that match category X but no types are configured (e.g. “Chemotherapy”) (so in theory it could be a PASS if types were known).

This is bug fix for LKO151 I-2.

Note that radiotherapies and surgeries can be ignored (this rule is about systemic treatment only)