ACTIN-1817 Use combined patient id + ihc to ignore LLM mistakes
Instead of using the free text, we can allow for a combo of patient id with IHC string
to match at a finer grain when the input is erroneous, without impacting future patients.

Curating against only the IHC value is also possible, but ignore will take precedence on a
per patient level.