ACTIN-1817 Use patient id + ihc to ignore/curate specific patient test result (#1050)
Instead of using the free text, we can allow for a combo of patient id with IHC string
to match at a finer grain when the input is erroneous, without impacting future patients.

Curating against only the IHC value is also possible, but ignore will take precedence on a
per patient level.