ACTIN-1596: Treatment ranking algorithm based on evidence scoring (#1105)
Add the scoring algorithm proposed by the MCGI team which uses tumor match, variant match and approval status to create a score per evidence. This score is summed per treatment across molecular events.

"Duplicate" evidences are summed with diminishing returns. A piece of evidence is considered duplicated when the event is in the same gene, matched on the same cancer type, and has the same direction.

The model is in the new interpretation module and is run in the report. There is a really basic table in the report which I would love some ideas on how to better flesh out.

See ticket for the details on the scoring.