ACTIN-2087: Create new message type which allows messages to be combined (#1185)
Makes the messages in evalutians into instances of EvaluationMessage, an interface with two functions that essentially work together to allow messages to be combined. The first says what messages should be combined, and the second how to combine them. It is up to the message creator how to define this behaviour.

The StaticMessage is a type of EvaluationMessage that behaves as messages do now. The only messages that can be "combined" are exact duplicates, and in this case the combination is just to remove one of the two.

The TargetCoverageMessage shows what we can do now. It maintains a list of genes (dynamic) and message string (static) in the state of the object. It can be "combined" when the static parts of the message are equal, and the combination operation concatenates the list of genes.

We can introduce other subclasses of EvaluationMessage whenever we need some non-trivial way to combine like messages together. And could also add to the interface (for instance include the evaluation class type) to implement things like ACTIN-13