ACTIN-2552: Indirect evidence matching for hotspots (#1319)
Efficacy evidence now includes indirect matches, which are variants on the same gene
with the same protein effect as a patient variant, for treatments that are generic inhibitors
and for which the variant is not associated with resistance.

MolecularMatchDetails now has an additional boolean indicating if a match is indirect.

---------

Co-authored-by: Korneel Duyvesteyn <k.duyvesteyn@hartwigmedicalfoundation.nl>